+++ trace-protobuf/build.rs
@@ -61,59 +61,59 @@ fn generate_protobuf() {
     config.type_attribute("Span", "#[derive(Deserialize, Serialize)]");
     config.field_attribute(
         ".pb.Span.service",
-        "#[serde(default)] #[serde(deserialize_with = \"crate::serializers::deserialize_null_into_default\")]",
+        "#[serde(default)] #[serde(deserialize_with = \"crate::deserializers::deserialize_null_into_default\")]",
     );
     config.field_attribute(
         ".pb.Span.name",
-        "#[serde(default)] #[serde(deserialize_with = \"crate::serializers::deserialize_null_into_default\")]",
+        "#[serde(default)] #[serde(deserialize_with = \"crate::deserializers::deserialize_null_into_default\")]",
     );
     config.field_attribute(
         ".pb.Span.resource",
-        "#[serde(default)] #[serde(deserialize_with = \"crate::serializers::deserialize_null_into_default\")]",
+        "#[serde(default)] #[serde(deserialize_with = \"crate::deserializers::deserialize_null_into_default\")]",
     );
     config.field_attribute(
         ".pb.Span.traceID",
-        "#[serde(default)] #[serde(deserialize_with = \"crate::serializers::deserialize_null_into_default\")]",
+        "#[serde(default)] #[serde(deserialize_with = \"crate::deserializers::deserialize_null_into_default\")]",
     );
     config.field_attribute(
         ".pb.Span.spanID",
-        "#[serde(default)] #[serde(deserialize_with = \"crate::serializers::deserialize_null_into_default\")]",
+        "#[serde(default)] #[serde(deserialize_with = \"crate::deserializers::deserialize_null_into_default\")]",
     );
     config.field_attribute(
         ".pb.Span.parentID",
-        "#[serde(default)] #[serde(deserialize_with = \"crate::serializers::deserialize_null_into_default\")]",
+        "#[serde(default)] #[serde(deserialize_with = \"crate::deserializers::deserialize_null_into_default\")]",
     );
     config.field_attribute(
         ".pb.Span.start",
-        "#[serde(default)] #[serde(deserialize_with = \"crate::serializers::deserialize_null_into_default\")]",
+        "#[serde(default)] #[serde(deserialize_with = \"crate::deserializers::deserialize_null_into_default\")]",
     );
     config.field_attribute(
         ".pb.Span.duration",
-        "#[serde(default)] #[serde(deserialize_with = \"crate::serializers::deserialize_duration\")]",
+        "#[serde(default)] #[serde(deserialize_with = \"crate::deserializers::deserialize_duration\")]",
     );
     config.field_attribute(
         ".pb.Span.meta",
-        "#[serde(default)] #[serde(deserialize_with = \"crate::serializers::deserialize_null_into_default\")]",
+        "#[serde(default)] #[serde(deserialize_with = \"crate::deserializers::deserialize_null_into_default\")]",
     );
     config.field_attribute(
         ".pb.Span.metrics",
-        "#[serde(default)] #[serde(deserialize_with = \"crate::serializers::deserialize_null_into_default\")]",
+        "#[serde(default)] #[serde(deserialize_with = \"crate::deserializers::deserialize_null_into_default\")]",
     );
     config.field_attribute(
         ".pb.Span.type",
-        "#[serde(default)] #[serde(deserialize_with = \"crate::serializers::deserialize_null_into_default\")]",
+        "#[serde(default)] #[serde(deserialize_with = \"crate::deserializers::deserialize_null_into_default\")]",
     );
     config.field_attribute(
         ".pb.Span.meta_struct",
-        "#[serde(default)] #[serde(deserialize_with = \"crate::serializers::deserialize_null_into_default\")] #[serde(skip_serializing_if = \"::std::collections::HashMap::is_empty\")]",
+        "#[serde(default)] #[serde(deserialize_with = \"crate::deserializers::deserialize_null_into_default\")] #[serde(skip_serializing_if = \"::std::collections::HashMap::is_empty\")]",
     );
     config.field_attribute(
         ".pb.Span.spanLinks",
-        "#[serde(default)] #[serde(deserialize_with = \"crate::serializers::deserialize_null_into_default\")] #[serde(skip_serializing_if = \"::prost::alloc::vec::Vec::is_empty\")]",
+        "#[serde(default)] #[serde(deserialize_with = \"crate::deserializers::deserialize_null_into_default\")] #[serde(skip_serializing_if = \"::prost::alloc::vec::Vec::is_empty\")]",
     );
     config.field_attribute(
         ".pb.Span.error",
-        "#[serde(default)] #[serde(deserialize_with = \"crate::serializers::deserialize_null_into_default\")] #[serde(skip_serializing_if = \"crate::serializers::is_default\")]",
+        "#[serde(default)] #[serde(deserialize_with = \"crate::deserializers::deserialize_null_into_default\")] #[serde(skip_serializing_if = \"crate::deserializers::is_default\")]",
     );
 
     config.type_attribute("StatsPayload", "#[derive(Deserialize, Serialize)]");
